+++ pytket/extensions/qiskit/backends/ibmq_emulator.py
@@ -54,10 +54,10 @@ from .ibm_utils import _STATUS_MAP, _batch_circuits
 
 
 class IBMQEmulatorBackend(Backend):
-    """A Backend which uses the AerBackend simulator to emulate the behaviour of
+    """A Backend which uses the ibmq_qasm_simulator to emulate the behaviour of
     IBMQBackend. Identical to :py:class:`IBMQBackend` except there is no `monitor`
     parameter. Performs the same compilation and predicate checks as IBMQBackend.
-    Requires a valid IBM account.
+    Requires a valid IBMQ account.
     """
 
     _supports_shots = False