+++ docs/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 ~~~~~~~~~
 
+Unreleased
+----------
+
+* Update pytket version requirement to 1.25.
+* Update qiskit version requirement to 0.46.
+* Update qiskit-ibm-provider version requirement to 0.10.
+* Update qiskit-ibm-runtime version requirement to 0.19.
+
 0.48.1rc1
 ---------