+++ docs/changelog.rst
@@ -4,6 +4,10 @@ Changelog
 Unreleased
 ----------
 
+* Update pytket version requirement to 1.25.
+* Update qiskit version requirement to 0.46.
+* Update qiskit-ibm-provider version requirement to 0.10.
+* Update qiskit-ibm-runtime version requirement to 0.19.
 * Add ``IBMQLocalEmulatorBackend`` for running local emulation of
   ``IBMQBackend`` using ``AerBackend`` with a noise model.