+++ docs/changelog.rst
@@ -5,9 +5,9 @@ Unreleased
 ----------
 
 * Update pytket version requirement to 1.25.
-* Update qiskit version requirement to 0.46.
+* Update qiskit version requirement to 1.0.
 * Update qiskit-ibm-provider version requirement to 0.10.
-* Update qiskit-ibm-runtime version requirement to 0.19.
+* Update qiskit-ibm-runtime version requirement to 0.20.
 * Add ``IBMQLocalEmulatorBackend`` for running local emulation of
   ``IBMQBackend`` using ``AerBackend`` with a noise model.