+++ docs/changelog.rst
@@ -7,7 +7,7 @@ Unreleased
 * Update pytket version requirement to 1.25.
 * Update qiskit version requirement to 1.0.
 * Update qiskit-ibm-provider version requirement to 0.10.
-* Update qiskit-ibm-runtime version requirement to 0.20.
+* Update qiskit-ibm-runtime version requirement to 0.19.
 * Add ``IBMQLocalEmulatorBackend`` for running local emulation of
   ``IBMQBackend`` using ``AerBackend`` with a noise model.